refactor: Move scraping function to services dir
